% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linechart.R
\name{linechart}
\alias{linechart}
\title{Using a dataframe specific column to plot a line figure}
\usage{
linechart(da, t, a, b, c, d)
}
\arguments{
\item{data}{a dataframe or tibble}
}
\value{
a figure
}
\description{
Using a dataframe specific column to plot a line figure
}
\examples{
linechart(data,t,a,b,c,d)
}
